added a getFileNameFromPath function
NOTE: this will be moved into my skele_lib in the future but for now it is stand alone.
this makes it so that you can get the file name given a file path. this also works for getting the bottom level folder that you're in.

added str_len function this is made to get the length of a cstring ([const]char*) this is the right way to do it. will break if not a null terminated c_string

I used getFileNameFromPath in the main function to give usage hints if you don't pass in 2 pramas this will eventually be handled by  Skele_lib::UsageLib